fix(shared): validate condition webhook URLs against SSRF
Extract validateUrlSsrf to application-generic and reuse it for the
conditions filter webhook and the HTTP request workflow step.
