+++ libs/application-generic/src/usecases/conditions-filter/conditions-filter.usecase.ts
@@ -30,7 +30,14 @@ import axios from 'axios';
 import { differenceInDays, differenceInHours, differenceInMinutes, parseISO } from 'date-fns';
 import { decryptApiKey } from '../../encryption';
 import { buildSubscriberKey, CachedResponse } from '../../services';
-import { createHash, Filter, FilterProcessingDetails, IFilterVariables, PlatformException } from '../../utils';
+import {
+  createHash,
+  Filter,
+  FilterProcessingDetails,
+  IFilterVariables,
+  PlatformException,
+  validateUrlSsrf,
+} from '../../utils';
 import { CompileTemplate } from '../compile-template';
 import { CreateExecutionDetails, CreateExecutionDetailsCommand, DetailEnum } from '../create-execution-details';
 import { ConditionsFilterCommand } from './conditions-filter.command';
@@ -255,6 +262,17 @@ export class ConditionsFilter extends Filter {
 
     if (hmac) {
       config.headers['nv-hmac-256'] = hmac;
+    }
+
+    const ssrfError = await validateUrlSsrf(child.webhookUrl);
+
+    if (ssrfError) {
+      throw new Error(
+        JSON.stringify({
+          message: ssrfError,
+          data: 'Webhook URL blocked by SSRF protection.',
+        })
+      );
     }
 
     try {
