+++ libs/application-generic/src/utils/index.ts
@@ -28,6 +28,7 @@ export * from './object';
 export * from './parse-payload-schema';
 export * from './parse-step-variables';
 export * from './sanitize-control-values';
+export * from './ssrf-url-validation';
 export * from './step-resolver-control-state';
 export * from './step-type-to-control.mapper';
 export * from './subscriber';
